[#1599] Factor CustomDomain model back to single file
